ai-dental-receptionist/backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pydantic
google-auth
google-auth-oauthlib
google-api-python-client
faiss-cpu
sentence-transformers
numpy
